Some coding samples
K-means clustering parallelized with Pthreads in C++

Project for class which involved using a provided set of song lyrics to
identify songs which had a Jaccard similarity greater than .95.
Min-hashing was necessary due to the size of the data-set.

Python script showing an example of how I use Python in my day-to-day
work; this particular script looks at an atomic structure file (i.e.
list of atoms and coordinates) and creates a list of molecules based on
distances between atoms.